add swagger, fix liquibase, fix entity

+++ src/main/java/ru/spring/app/engine/repository/GlobalSettingsRepository.java
@@ -2,9 +2,9 @@ package ru.spring.app.engine.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import ru.spring.app.engine.model.GlobalSettings;
+import ru.spring.app.engine.entity.GlobalSettings;
 
 @Repository
-public interface GlobalSettingsRepository extends JpaRepository<GlobalSettings, Integer> {
+public interface GlobalSettingsRepository extends JpaRepository<GlobalSettings, Long> {
 
 }
